Refactor checkquery and add it to docs
