--- docs/source/scribe_data/checkquery.rst
@@ -1,8 +1,0 @@
-checkquery
-==========
-
-`View code on Github <https://github.com/scribe-org/Scribe-Data/blob/main/src/scribe_data/checkquery.py>`_
-
-.. automodule:: scribe_data.checkquery
-    :members:
-    :private-members:
