+++ docs/source/scribe_data/checkquery/check.rst
@@ -1,0 +1,8 @@
+check
+=====
+
+`View code on Github <https://github.com/scribe-org/Scribe-Data/blob/main/src/scribe_data/checkquery/check.py>`_
+
+.. automodule:: scribe_data.checkquery.check
+    :members:
+    :private-members:
