+++ docs/source/scribe_data/checkquery/index.rst
@@ -1,0 +1,11 @@
+checkquery
+==========
+
+`View code on Github <https://github.com/scribe-org/Scribe-Data/tree/main/src/scribe_data/checkquery>`_
+
+.. toctree::
+    :maxdepth: 1
+
+    check
+    query
+    sparql
