+++ docs/source/scribe_data/checkquery/query.rst
@@ -1,0 +1,8 @@
+query
+=====
+
+`View code on Github <https://github.com/scribe-org/Scribe-Data/blob/main/src/scribe_data/checkquery/query.py>`_
+
+.. automodule:: scribe_data.checkquery.query
+    :members:
+    :private-members:
