+++ docs/source/scribe_data/checkquery/sparql.rst
@@ -1,0 +1,8 @@
+sparql
+======
+
+`View code on Github <https://github.com/scribe-org/Scribe-Data/blob/main/src/scribe_data/checkquery/sparql.py>`_
+
+.. automodule:: scribe_data.checkquery.sparql
+    :members:
+    :private-members:
