+++ docs/source/scribe_data/extract_transform/unicode/emoji_utils.rst
@@ -1,8 +1,8 @@
-emoji_utils
-===========
+unicode_utils
+=============
 
-`View code on Github <https://github.com/scribe-org/Scribe-Data/blob/main/src/scribe_data/extract_transform/unicode/emoji_utils.py>`_
+`View code on Github <https://github.com/scribe-org/Scribe-Data/blob/main/src/scribe_data/extract_transform/unicode/unicode_utils.py>`_
 
-.. automodule:: scribe_data.extract_transform.unicode.emoji_utils
+.. automodule:: scribe_data.extract_transform.unicode.unicode_utils
     :members:
     :private-members:
