+++ src/scribe_data/checkquery/check.py
@@ -1,44 +1,25 @@
+"""
+Command line tool for testing SPARQl queries against an endpoint.
+"""
+
 import argparse
 import contextlib
 import os
 import pathlib
+import subprocess
 import sys
-import time
 import urllib.request
-from dataclasses import dataclass
-from typing import Optional, List, Tuple
+from typing import List, Optional, Tuple
 
-import SPARQLWrapper as SPARQL
-from SPARQLWrapper import SPARQLExceptions
 from tqdm.auto import tqdm
+
+from scribe_data.checkquery.query import QueryExecutionException, QueryFile
+from scribe_data.checkquery.sparql import execute, sparql_context
 
 EXIT_SUCCESS = 0
 EXIT_FAILURE = 1
-EXIT_CLI_ERROR = 2
 
 PROJECT_ROOT = "Scribe-Data"
-
-
-@dataclass(frozen=True)
-class QueryFile:
-    path: pathlib.Path
-
-    def load(self, limit: int) -> str:
-        with open(self.path, encoding="utf-8") as in_stream:
-            return f"{in_stream.read()}\nLIMIT {limit}\n"
-
-    def __repr__(self) -> str:
-        return f"{self.__class__.__name__}(path={self.path})"
-
-
-class QueryExecutionException(Exception):
-    def __init__(self, message: str, query: QueryFile) -> None:
-        self.message = message
-        self.query = query
-        super().__init__(self.message)
-
-    def __str__(self) -> str:
-        return f"{self.query.path} : {self.message}"
 
 
 def ping(url: str, timeout: int) -> bool:
@@ -86,31 +67,6 @@ def changed_queries() -> Optional[List[QueryFile]]:
     ]
 
     return [QueryFile(fpath) for fpath in changed_files if fpath.suffix == ".sparql"]
-
-
-def sparql_context(url: str) -> SPARQL.SPARQLWrapper:
-    context = SPARQL.SPARQLWrapper(url)
-    context.setReturnFormat(SPARQL.JSON)
-    context.setMethod(SPARQL.POST)
-    return context
-
-
-def execute(query: QueryFile, limit: int, context: SPARQL.SPARQLWrapper) -> dict:
-    def delay_in_seconds() -> int:
-        return int(math.ceil(10.0 / math.sqrt(3)))
-
-    try:
-        context.setQuery(query.load(limit))
-        return context.queryAndConvert()
-    except HTTPError:
-        time.sleep(delay_in_seconds())
-        return execute(query, limit, context)
-    except SPARQLExceptions.SPARQLWrapperException as err:
-        raise QueryExecutionException(err.msg, query) from err
-    except Exception as err:
-        raise QueryExecutionException(
-            f"{type(err).__name__} - {str(err)}", query
-        ) from err
 
 
 def check_sparql_file(fpath: str) -> pathlib.Path:
