+++ src/scribe_data/checkquery/query.py
@@ -1,3 +1,7 @@
+"""
+Classes and methods for querying a file in the query check process.
+"""
+
 import pathlib
 from dataclasses import dataclass
 
@@ -12,3 +16,13 @@ class QueryFile:
 
     def __repr__(self) -> str:
         return f"{self.__class__.__name__}(path={self.path})"
+
+
+class QueryExecutionException(Exception):
+    def __init__(self, message: str, query: QueryFile) -> None:
+        self.message = message
+        self.query = query
+        super().__init__(self.message)
+
+    def __str__(self) -> str:
+        return f"{self.query.path} : {self.message}"
