--- src/scribe_data/checkquery_utills/constants.py
@@ -1,5 +1,0 @@
-EXIT_SUCCESS = 0
-EXIT_FAILURE = 1
-EXIT_CLI_ERROR = 2
-
-PROJECT_ROOT = "Scribe-Data"
