+++ src/scribe_data/extract_transform/translation/translation_utils.py
@@ -1,9 +1,5 @@
 """
 Utility functions for the machine translation process.
-
-Contents:
-    translation_interrupt_handler,
-    translate_to_other_languages
 """
 
 import json
